# CGUITTFont optimizations