--- a/src/irrlicht_changes/CGUITTFont.cpp
+++ b/src/irrlicht_changes/CGUITTFont.cpp
@@ -562,6 +562,7 @@ void CGUITTFont::draw(const EnrichedString &text, const core::rect<s32>& positio
 	{
 		Glyph_Pages[i]->render_positions.clear();
 		Glyph_Pages[i]->render_source_rects.clear();
+		Glyph_Pages[i]->render_colors.clear();
 	}
 
 	// Set up some variables.
@@ -589,8 +590,6 @@ void CGUITTFont::draw(const EnrichedString &text, const core::rect<s32>& positio
 	// Start parsing characters.
 	uchar32_t previousChar = 0;
 	core::ustring::const_iterator iter(utext);
-	std::vector<video::SColor> applied_colors;
-	applied_colors.reserve(utext.size_raw());
 	while (!iter.atEnd())
 	{
 		uchar32_t currentChar = *iter;
@@ -636,10 +635,11 @@ void CGUITTFont::draw(const EnrichedString &text, const core::rect<s32>& positio
 			CGUITTGlyphPage* const page = Glyph_Pages[glyph.glyph_page];
 			page->render_positions.push_back(core::position2di(offset.X + offx, offset.Y + offy));
 			page->render_source_rects.push_back(glyph.source_rect);
+			if (iter.getPos() < colors.size())
+				page->render_colors.push_back(colors[iter.getPos()]);
+			else
+				page->render_colors.push_back(video::SColor(255,255,255,255));
 			Render_Map.set(glyph.glyph_page, page);
-			u32 current_color = iter.getPos();
-			if (current_color < colors.size())
-				applied_colors.push_back(colors[current_color]);
 		}
 		if (n > 0)
 		{
@@ -669,8 +669,6 @@ void CGUITTFont::draw(const EnrichedString &text, const core::rect<s32>& positio
 		previousChar = currentChar;
 		++iter;
 	}
-	for (u32 i = applied_colors.size(); i < utext.size_raw(); i++)
-		applied_colors.emplace_back(video::SColor(255, 255, 255, 255));
 
 	// Draw now.
 	update_glyph_pages();
@@ -695,10 +693,10 @@ void CGUITTFont::draw(const EnrichedString &text, const core::rect<s32>& positio
 		video::SColor colprev;
 		for (size_t i = 0; i < page->render_positions.size(); ++i) {
 			ibegin = i;
-			colprev = applied_colors[i];
+			colprev = page->render_colors[i];
 			do
 				++i;
-			while (i < page->render_positions.size() && applied_colors[i] == colprev);
+			while (i < page->render_positions.size() && page->render_colors[i] == colprev);
 			core::array<core::vector2di> tmp_positions;
 			core::array<core::recti> tmp_source_rects;
 			tmp_positions.set_pointer(&page->render_positions[ibegin], i - ibegin, false, false); // no copy