uses batch drawing instead of each glyph individually (as permitted by coloring)

also fixed coloring being messed up if multiple glyph pages are used, since this is a rare case nobody has noticed this for years

## To do

This PR is Ready for Review.