@Desour 2021-03-31
Tested. Chat drawing is faster.
Chat off: 60 fps, 18 ms update frame time
Chat on, filled with multiple invocations of `/mods`:
master: 19 fps, 52 ms update frame time
this PR: 33 fps, 35 ms update frame time